clean up bi-ags
