+++ data/paper/gs_arch.json
@@ -1,0 +1,1 @@
+{"compas": "0.17.2", "datatype": "compas_ags.diagrams/FormGraph", "data": {"attributes": {"name": "Network"}, "node_attributes": {"x": 0.0, "y": 0.0, "z": 0.0}, "edge_attributes": {}, "node": {"0": {"x": 3.62477467512, "y": 2.99447312681, "z": 0.0}, "1": {"x": 43.4972961014, "y": -7.27758178128, "z": 0.0}, "2": {"x": 0.0, "y": 0.0, "z": 0.0}, "3": {"x": 0.0, "y": -7.27758178128, "z": 0.0}, "4": {"x": 39.8725214263, "y": 2.99447312681, "z": 0.0}, "5": {"x": 32.622972076, "y": 6.81140730234, "z": 0.0}, "6": {"x": 25.3734227258, "y": 8.54514654776, "z": 0.0}, "7": {"x": 18.1238733756, "y": 8.54514654776, "z": 0.0}, "8": {"x": 10.8743240253, "y": 6.81140730234, "z": 0.0}, "9": {"x": 43.4972961014, "y": 0.0, "z": 0.0}, "10": {"x": 3.62477467512, "y": 10.2720549081, "z": 0.0}, "11": {"x": -3.52953624469, "y": 0.0, "z": 0.0}, "12": {"x": 39.8725214263, "y": 10.2720549081, "z": 0.0}, "13": {"x": 32.622972076, "y": 14.0889890836, "z": 0.0}, "14": {"x": 25.3734227258, "y": 15.822728329, "z": 0.0}, "15": {"x": 18.1238733756, "y": 15.822728329, "z": 0.0}, "16": {"x": 10.8743240253, "y": 14.0889890836, "z": 0.0}, "17": {"x": 47.4502140636, "y": 0.0, "z": 0.0}}, "edge": {"0": {"10": {}, "8": {}}, "1": {"9": {}}, "2": {"11": {}, "0": {}}, "3": {"2": {}}, "4": {"12": {}, "9": {}}, "5": {"13": {}, "4": {}}, "6": {"14": {}, "5": {}}, "7": {"15": {}, "6": {}}, "8": {"16": {}, "7": {}}, "9": {"17": {}}, "10": {}, "11": {}, "12": {}, "13": {}, "14": {}, "15": {}, "16": {}, "17": {}}, "adjacency": {"0": {"10": null, "2": null, "8": null}, "1": {"9": null}, "2": {"11": null, "3": null, "0": null}, "3": {"2": null}, "4": {"12": null, "5": null, "9": null}, "5": {"13": null, "6": null, "4": null}, "6": {"14": null, "7": null, "5": null}, "7": {"15": null, "8": null, "6": null}, "8": {"16": null, "0": null, "7": null}, "9": {"1": null, "4": null, "17": null}, "10": {"0": null}, "11": {"2": null}, "12": {"4": null}, "13": {"5": null}, "14": {"6": null}, "15": {"7": null}, "16": {"8": null}, "17": {"9": null}}, "max_int_key": 17}}
