+++ data/paper/gs_arch_2layers.json
@@ -1,0 +1,1 @@
+{"compas": "0.17.2", "datatype": "compas_ags.diagrams/FormGraph", "data": {"attributes": {"name": "Network"}, "node_attributes": {"x": 0.0, "y": 0.0, "z": 0.0}, "edge_attributes": {}, "node": {"0": {"x": 10.9336764457, "y": 2.93969333667, "z": 0.0}, "1": {"x": 4.27834823297, "y": 0.934076623109, "z": 0.0}, "2": {"x": 22.948337743, "y": 3.5619857547, "z": 0.0}, "3": {"x": 17.051662257, "y": 3.5619857547, "z": 0.0}, "4": {"x": 36.6666666667, "y": 2.75370967411, "z": 0.0}, "5": {"x": 30.0, "y": 6.26375238264, "z": 0.0}, "6": {"x": 40.0, "y": 0.0, "z": 0.0}, "7": {"x": 35.721651767, "y": 0.934076623109, "z": 0.0}, "8": {"x": 29.0663235543, "y": 2.93969333667, "z": 0.0}, "9": {"x": 0.0, "y": 0.0, "z": 0.0}, "10": {"x": 3.33333333333, "y": 2.75370967411, "z": 0.0}, "11": {"x": 10.0, "y": 6.26375238264, "z": 0.0}, "12": {"x": 16.6666666667, "y": 7.85809446899, "z": 0.0}, "13": {"x": 23.3333333333, "y": 7.85809446899, "z": 0.0}, "14": {"x": 0.0, "y": -6.69244521712, "z": 0.0}, "15": {"x": 40.0, "y": -6.69244521712, "z": 0.0}, "16": {"x": 3.33333333333, "y": 9.44615489123, "z": 0.0}, "17": {"x": 10.0, "y": 12.9561975998, "z": 0.0}, "18": {"x": 16.6666666667, "y": 14.5505396861, "z": 0.0}, "19": {"x": 23.3333333333, "y": 14.5505396861, "z": 0.0}, "20": {"x": 30.0, "y": 12.9561975998, "z": 0.0}, "21": {"x": 36.6666666667, "y": 9.44615489123, "z": 0.0}, "22": {"x": -3.24575232121, "y": 0.0, "z": 0.0}, "23": {"x": 43.6350930439, "y": 0.0, "z": 0.0}}, "edge": {"0": {"11": {}, "3": {}}, "1": {"10": {}, "0": {}}, "2": {"13": {}}, "3": {"12": {}, "2": {}}, "4": {"7": {}, "21": {}, "6": {}}, "5": {"8": {}, "20": {}, "4": {}}, "6": {"7": {}, "23": {}}, "7": {"8": {}}, "8": {"2": {}}, "9": {"1": {}, "22": {}, "10": {}}, "10": {"16": {}, "11": {}}, "11": {"17": {}, "12": {}}, "12": {"18": {}, "13": {}}, "13": {"19": {}, "5": {}}, "14": {"9": {}}, "15": {"6": {}}, "16": {}, "17": {}, "18": {}, "19": {}, "20": {}, "21": {}, "22": {}, "23": {}}, "adjacency": {"0": {"11": null, "3": null, "1": null}, "1": {"10": null, "0": null, "9": null}, "2": {"13": null, "3": null, "8": null}, "3": {"12": null, "2": null, "0": null}, "4": {"7": null, "21": null, "5": null, "6": null}, "5": {"8": null, "20": null, "13": null, "4": null}, "6": {"7": null, "15": null, "4": null, "23": null}, "7": {"4": null, "6": null, "8": null}, "8": {"5": null, "7": null, "2": null}, "9": {"1": null, "14": null, "22": null, "10": null}, "10": {"1": null, "16": null, "9": null, "11": null}, "11": {"0": null, "17": null, "10": null, "12": null}, "12": {"3": null, "18": null, "11": null, "13": null}, "13": {"2": null, "19": null, "12": null, "5": null}, "14": {"9": null}, "15": {"6": null}, "16": {"10": null}, "17": {"11": null}, "18": {"12": null}, "19": {"13": null}, "20": {"5": null}, "21": {"4": null}, "22": {"9": null}, "23": {"6": null}}, "max_int_key": 23}}
