+++ examples/bi_ags/example_rf_3-1_nullspace.py
@@ -23,16 +23,15 @@ from compas_ags.ags import ConstraintsCollection, HorizontalFix, VerticalFix, An
 # ------------------------------------------------------------------------------
 #   1. get lines of a plane triangle frame in equilibrium, its applied loads and reaction forces
 # ------------------------------------------------------------------------------
-graph = FormGraph.from_obj(compas_ags.get('paper/gs_form_force.obj'))
 
+graph = FormGraph.from_obj(compas_ags.get('paper/gs_form_force.obj'))
 form = FormDiagram.from_graph(graph)
 force = ForceDiagram.from_formdiagram(form)
-
 
 # ------------------------------------------------------------------------------
 #   2. set the fixed vertices
 # ------------------------------------------------------------------------------
-left = list(form.vertices_where({'x': 0.0, 'y': 0.0}))[0]
+left  = list(form.vertices_where({'x': 0.0, 'y': 0.0}))[0]
 right = list(form.vertices_where({'x': 6.0, 'y': 0.0}))[0]
 fixed = [left, right]
 
@@ -43,8 +42,8 @@ form.vertex_attribute(key, 'is_anchor', True)
 # ------------------------------------------------------------------------------
 #   3. set applied load
 # ------------------------------------------------------------------------------
-e1 = {'v': list(form.vertices_where({'x': 3.0, 'y': 3.0}))[0],
-      'u': list(form.vertices_where({'x': 3.669563106796117, 'y': 5.008689320388349}))[0]}
+e1 ={'v': list(form.vertices_where({'x': 3.0, 'y': 3.0}))[0],
+     'u': list(form.vertices_where({'x': 3.669563106796117, 'y': 5.008689320388349}))[0]}
 form.edge_attribute((e1['v'], e1['u']), 'q', -1.0)
 form.edge_attribute((e1['v'], e1['u']), 'is_ind', True)
 
@@ -55,15 +54,16 @@ graphstatics.force_update_from_form(force, form)
 # store the original vertex locations
 force_key_xyz = {key: force.vertex_coordinates(key) for key in force.vertices()}
 
+
 # --------------------------------------------------------------------------
 #   4. force diagram manipulation and ompute the nullspace
 # --------------------------------------------------------------------------
 # set constraints
 C = ConstraintsCollection(form)
 
-# fix y coordinates of the left and right vertices
-C.add_constraint(HorizontalFix(form, left))
-C.add_constraint(HorizontalFix(form, right))
+# fix an angle to left and right vertices
+C.add_constraint(AngleFix(form, left, 30))
+C.add_constraint(AngleFix(form, right, 30))
 
 # fix the length of edges connecting leaf vertices
 C.constrain_dependent_leaf_edges_lengths()
@@ -73,6 +73,7 @@ constraint_lines = C.get_lines()
 # which means the amount of independent solutions of form diagrams
 ns = compute_nullspace(form, force, C)
 print("Dimension of nullspace: " + str(len(ns)))
+
 
 # --------------------------------------------------------------------------
 #   5. display force diagram and a specific solution of form diagram
@@ -87,11 +88,12 @@ def show(i):
     for u, v in form.edges():
         form_lines.append({
             'start': [x + y for x, y in zip(form.vertex_coordinates(u, 'xy'),  nsi[u])],
-            'end': [x + y for x, y in zip(form.vertex_coordinates(v, 'xy'),  nsi[v])],
+            'end'  : [x + y for x, y in zip(form.vertex_coordinates(v, 'xy'),  nsi[v])],
             'width': 1.0,
             'color': '#cccccc',
             'style': '--'
         })
+
 
     form_lines = form_lines + constraint_lines
 
@@ -105,14 +107,13 @@ def show(i):
                      vertexsize=0.2,
                      vertexcolor={key: '#000000' for key in fixed},
                      edgelabel={uv: index for index, uv in enumerate(form.edges())}
-                     )
+    )
 
     viewer.draw_force(vertexlabel={key: key for key in force.vertices()},
                       vertexsize=0.2,
                       edgelabel={uv: index for index, uv in enumerate(force.edges())}
-                      )
+    )
 
     viewer.show()
-
 
 show(0)
